Fix delete of programs (include programs_users).

+++ routes/security/functional_areas.rb
@@ -103,6 +103,7 @@ class Framework < Roda
         end
         r.delete do    # DELETE
           return_json_response
+          raise Crossbeams::AuthorizationError unless authorised?('menu', 'delete')
           res = interactor.delete_program(id)
           flash[:notice] = res.message
           redirect_to_last_grid(r)
@@ -188,6 +189,7 @@ class Framework < Roda
         end
         r.delete do    # DELETE
           return_json_response
+          raise Crossbeams::AuthorizationError unless authorised?('menu', 'delete')
           res = interactor.delete_program_function(id)
           flash[:notice] = res.message
           redirect_to_last_grid(r)
